Separated out components, remove some more stuff, update styling a little, fix wallet connect button not showing up
