+++ packages/react-app/src/components/Wallet.jsx
@@ -1,0 +1,47 @@
+import { Account } from "./index";
+import React, { useCallback, useEffect } from "react";
+import { ethers } from "ethers";
+import WalletSetup from "../helpers/WalletSetup";
+
+export default function Wallet({ injectedProvider, setInjectedProvider }) {
+  const web3Modal = WalletSetup();
+
+  const logoutOfWeb3Modal = async () => {
+    await web3Modal.clearCachedProvider();
+    if (injectedProvider && injectedProvider.provider && typeof injectedProvider.provider.disconnect == "function") {
+      await injectedProvider.provider.disconnect();
+    }
+    setTimeout(() => {
+      window.location.reload();
+    }, 1);
+  };
+
+  const loadWeb3Modal = useCallback(async () => {
+    const provider = await web3Modal.connect();
+    setInjectedProvider(new ethers.providers.Web3Provider(provider));
+
+    provider.on("chainChanged", () => {
+      setInjectedProvider(new ethers.providers.Web3Provider(provider));
+    });
+
+    provider.on("accountsChanged", () => {
+      setInjectedProvider(new ethers.providers.Web3Provider(provider));
+    });
+
+    provider.on("disconnect", () => {
+      logoutOfWeb3Modal();
+    });
+  }, [setInjectedProvider]);
+
+  useEffect(() => {
+    if (web3Modal.cachedProvider) {
+      loadWeb3Modal();
+    }
+  }, [loadWeb3Modal]);
+
+  return (
+    <div style={{ position: "fixed", textAlign: "right", right: 0, top: 0, padding: 5 }}>
+      <Account web3Modal={web3Modal} loadWeb3Modal={loadWeb3Modal} logoutOfWeb3Modal={logoutOfWeb3Modal} />
+    </div>
+  );
+}
