+++ packages/react-app/src/components/index.js
@@ -1,6 +1,9 @@
 export { default as Account } from "./Account";
 export { default as Balance } from "./Balance";
-export { default as BytesStringInput } from "./BytesStringInput";
 export { default as Contract } from "./Contract";
-export { default as HeaderBar } from "./HeaderBar";
+export { default as NavBar } from "./NavBar";
 export { default as Provider } from "./Provider";
+export { default as Staking } from "./Staking";
+export { default as NetworkDisplay } from "./NetworkDisplay";
+export { default as Yield } from "./Yield";
+export { default as CrownBalance } from "./CrownBalance";
